@@ -2,7 +2,7 @@
     "name": "magento/updater",
     "description": "Magento Updater",
     "type": "project",
-    "version": "10.0.3",
+    "version": "10.1.0",
     "license": [
         "OSL-3.0",
         "AFL-3.0"
@@ -14,15 +14,15 @@
         }
     ],
     "require": {
-        "php": "~5.5.0|~5.6.0|~7.0.0",
-        "composer/composer": "1.0.0-beta1",
-        "magento/composer": "~1.0.0",
-        "symfony/console": "~2.3 <2.7",
-        "monolog/monolog": "1.16.0",
+        "php": "~7.1.3||~7.2.0",
+        "composer/composer": "^1.6",
+        "magento/composer": "~1.3.0",
+        "symfony/console": "~4.0.0",
+        "monolog/monolog": "^1.17",
         "psr/log": "~1.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "4.1.0"
+        "phpunit/phpunit": "~7.0.0"
     },
     "autoload": {
         "psr-4": {
