@@ -1,11 +1,11 @@
 <?php
 /**
- * Copyright © 2016 Magento. All rights reserved.
+ * Copyright © 2013-2017 Magento, Inc. All rights reserved.
  * See COPYING.txt for license details.
  */
 namespace Magento\Update;
 
-class StatusCheckTest extends \PHPUnit_Framework_TestCase
+class StatusCheckTest extends \PHPUnit\Framework\TestCase
 {
     const REQUEST_TYPE_AJAX = 'ajax';
 
